+++ cursive/src/cursive_runnable.rs
@@ -1,4 +1,4 @@
-use crate::{backend, backends, Cursive};
+use crate::{backend, backends, Cursive, CursiveRunner};
 
 type Initializer =
     dyn FnMut()
@@ -61,6 +61,58 @@ impl CursiveRunnable {
     /// Runs the event loop with the registered backend initializer.
     pub fn try_run(&mut self) -> Result<(), Box<dyn std::error::Error>> {
         self.siv.try_run_with(&mut self.backend_init)
+    }
+
+    /// Gets a runner with the registered backend.
+    ///
+    /// Used to manually control the event loop. In most cases, running
+    /// `run()` will be easier.
+    ///
+    /// The runner will borrow `self`; when dropped, it will clear out the
+    /// terminal, and the cursive instance will be ready for another run if
+    /// needed.
+    pub fn try_runner(
+        &mut self,
+    ) -> Result<CursiveRunner<&mut Cursive>, Box<dyn std::error::Error>> {
+        Ok(self.siv.runner((self.backend_init)()?))
+    }
+
+    /// Gets a runner with the registered backend.
+    ///
+    /// # Panics
+    ///
+    /// If the backend initialization fails.
+    pub fn runner(&mut self) -> CursiveRunner<&mut Cursive> {
+        self.try_runner().unwrap()
+    }
+
+    /// Returns a new runner on the registered backend.
+    ///
+    /// Used to manually control the event loop. In most cases, running
+    /// `run()` will be easier.
+    ///
+    /// The runner will embed `self`; when dropped, it will clear out the
+    /// terminal, and the cursive instance will be dropped as well.
+    pub fn try_into_runner(
+        mut self,
+    ) -> Result<CursiveRunner<Self>, Box<dyn std::error::Error>> {
+        let backend = (self.backend_init)()?;
+        Ok(CursiveRunner::new(self, backend))
+    }
+
+    /// Returns a new runner on the registered backend.
+    ///
+    /// Used to manually control the event loop. In most cases, running
+    /// `run()` will be easier.
+    ///
+    /// The runner will embed `self`; when dropped, it will clear out the
+    /// terminal, and the cursive instance will be dropped as well.
+    ///
+    /// # Panics
+    ///
+    /// If the backend initialization fails.
+    pub fn into_runner(self) -> CursiveRunner<Self> {
+        self.try_into_runner().unwrap()
     }
 
     /// Creates a new Cursive wrapper using the dummy backend.